--- inventory-todollist-p/adapters/persistence/src/main/java/io/reflectoring/reviewapp/adapter/persistence/AuthorRepository.java
@@ -1,7 +1,0 @@
-package io.reflectoring.reviewapp.adapter.persistence;
-
-import io.reflectoring.reviewapp.domain.Author;
-import org.springframework.data.repository.CrudRepository;
-
-interface AuthorRepository extends CrudRepository<Author, Long> {
-}
